Finalize Quiz Master for submission

+++ logic/common/quiz_manager.py
@@ -10,6 +10,7 @@ from core.services.app_context import Services
 from data.quiz_repo import QuizRepository
 from logic.base_logic import BaseLogic
 from models.quiz import Quiz
+from ui.components.dialog import confirm_warning
 from ui.screens.common.quiz_manager import CommonQuizManagerScreen
 from utils.error_messages import QUIZ_ERROR_MESSAGES, format_errors
 
@@ -82,10 +83,7 @@ class CommonQuizManagerLogic(BaseLogic):
         self.invalid_quizzes.clear()
 
         quizzes = self.quiz_repo.get_all()
-
         self.quizzes = list(quizzes.values())
-        if not self.quizzes:
-            return
 
         valid_quizzes = []
 
@@ -122,10 +120,10 @@ class CommonQuizManagerLogic(BaseLogic):
         elif self.current_sort is QuizSortingOrder.OLDEST:
             custom_quizzes.sort(key=lambda q: q.updated_at)
             default_quizzes.sort(key=lambda q: q.quiz_title)
-        elif self.current_sort is QuizSortingOrder.NAME_ASC:
+        elif self.current_sort is QuizSortingOrder.TITLE_ASC:
             custom_quizzes.sort(key=lambda q: q.quiz_title)
             default_quizzes.sort(key=lambda q: q.quiz_title)
-        elif self.current_sort is QuizSortingOrder.NAME_DESC:
+        elif self.current_sort is QuizSortingOrder.TITLE_DESC:
             custom_quizzes.sort(key=lambda q: q.quiz_title, reverse=True)
             default_quizzes.sort(key=lambda q: q.quiz_title, reverse=True)
 
@@ -142,14 +140,34 @@ class CommonQuizManagerLogic(BaseLogic):
             self.custom_quizzes = len(custom_quizzes)
             self.screen.set_quizzes_number(self.total_quizzes, self.custom_quizzes)
 
+    def subtract_from_counter(self, include_custom: bool = True) -> None:
+        """
+        Subtracts one from the UI counters, used when removing a single quiz. If removing a default quiz, set
+        `include_custom` to False. Otherwise, set it to True.
+
+        Arguments:
+            include_custom: Whether to subtract one from the custom quizzes counter as well as the total quizzes
+                counter. Defaults to True.
+
+        Returns:
+            None.
+        """
+        self.total_quizzes -= 1
+
+        if include_custom:
+            self.custom_quizzes -= 1
+
+        self.screen.set_quizzes_number(self.total_quizzes, self.custom_quizzes)
+
     def _on_edit_requested(self, quiz: Quiz) -> None:
         """
         Internal method. Intended to be called when a quiz is requested to be edited by the user, or to be
         previewed if the quiz is a default quiz.
 
-        If the quiz is not a default quiz, opens the quiz setup screen to allow editing the quiz details
-        such as the title. Otherwise, the editor is opened directly if the quiz is a default quiz, as for
-        those quizzes, the Edit button acts instead as the Preview button.
+        If the quiz no longer exists, it is removed from the UI. If the quiz is not a default quiz, opens
+        the quiz setup screen to allow editing the quiz details such as the title. Otherwise, the editor is
+        opened directly if the quiz is a default quiz, as for those quizzes, the Edit button acts instead as
+        the Preview button.
 
         Arguments:
             quiz: The Quiz instance to edit/preview. A Quiz instance is used as it can be easily used to
@@ -159,6 +177,19 @@ class CommonQuizManagerLogic(BaseLogic):
             None.
         """
         self.quiz_repo.refresh_cache()
+
+        # Check if the quiz still exists
+        if not self.quiz_repo.exists(quiz.quiz_id):
+            self.screen.show_error(
+                "Quiz Not Found",
+                f"The quiz you're trying to {'preview' if quiz.is_premade else 'edit'} could not be found. It may have been deleted, or you may no longer have access to it.",
+            )
+
+            self.refresh_quizzes()
+
+            # Don't subtract from the custom counter if it's a default quiz
+            self.subtract_from_counter(include_custom=not quiz.is_premade)
+            return
 
         if not quiz.is_premade:
             # Opens setup screen to allow editing high-level quiz details
@@ -179,7 +210,7 @@ class CommonQuizManagerLogic(BaseLogic):
         Internal method. Intended to be called when a quiz is requested to be deleted by the user. Default
         quizzes cannot be deleted.
 
-        This method does not contain a confirmation prompt. It removes the quiz save file from disk and
+        If the quiz no longer exists, it is removed from the UI. It removes the quiz save file from disk and
         updates the UI to remove it from the UI as well. The counters are also updated to reflect the removal.
 
         Arguments:
@@ -191,6 +222,17 @@ class CommonQuizManagerLogic(BaseLogic):
         """
         self.quiz_repo.refresh_cache()
 
+        # Check if the quiz still exists
+        if not self.quiz_repo.exists(quiz.quiz_id):
+            self.screen.show_error(
+                "Quiz Not Found",
+                "The quiz you're trying to delete could not be found. It may have already been deleted.",
+            )
+
+            self.refresh_quizzes()
+            self.subtract_from_counter()
+            return
+
         if quiz.is_premade:
             # Should never happen in regular use; this is here as a safeguard
             self.screen.show_error(
@@ -199,14 +241,21 @@ class CommonQuizManagerLogic(BaseLogic):
             )
             return
 
+        # Confirm before deleting
+        confirm = confirm_warning(
+            self.screen,
+            "Confirm Deleting Quiz",
+            f'Are you sure you want to permanently delete the quiz "{quiz.quiz_title}"? This cannot be undone!',
+        )
+
+        if not confirm:
+            return
+
         # Remove from disk and refresh list to show that
         self.quiz_repo.remove(quiz.quiz_id)
-        self.refresh_quizzes()
 
-        # Update counters
-        self.total_quizzes -= 1
-        self.custom_quizzes -= 1
-        self.screen.set_quizzes_number(self.total_quizzes, self.custom_quizzes)
+        self.refresh_quizzes()
+        self.subtract_from_counter()
 
     def _on_invalid_quiz_info_requested(self) -> None:
         """
